health-first-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pymongo==4.6.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
loguru==0.7.2 

PyJWT==2.8.0
requests==2.31.0
pytz==2023.3 

PyJWT==2.8.0 

